% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_stocks_and_hedge_funds.R
\docType{package}
\name{BRStocksAndHedgeFunds}
\alias{BRStocksAndHedgeFunds}
\alias{BRStocksAndHedgeFunds-package}
\title{BRStocksAndHedgeFunds}
\description{
This package provides functions to download and process:
\itemize{
\item{Brazilian stocks prices and returns.}
\item{Brazilian stocks fundamentals.}
\item{Brazilian hedge funds prices and returns.}
}
}
\author{
Renato Dinhani \email{renatodinhani@gmail.com}
}
