% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_indicators.R
\name{download_indicator_quandl}
\alias{download_indicator_quandl}
\alias{download_indicator_cdi_daily}
\alias{download_indicator_cdi_monthly}
\alias{download_indicator_ibovespa_daily}
\alias{download_indicator_ipca_accumulated}
\alias{download_indicator_ipca_monthly}
\alias{download_indicator_selic_daily}
\alias{download_indicator_selic_monthly}
\alias{download_indicator_tjlp_monthly}
\title{Download Indicators}
\usage{
download_indicator_quandl(indicator_id, indicator_name = "",
  is_percentage = TRUE)

download_indicator_cdi_daily()

download_indicator_cdi_monthly()

download_indicator_ibovespa_daily()

download_indicator_ipca_accumulated()

download_indicator_ipca_monthly()

download_indicator_selic_daily()

download_indicator_selic_monthly()

download_indicator_tjlp_monthly()
}
\arguments{
\item{indicator_id}{Chracter. The Quandl unique ID of the desired indicator.}

\item{indicator_name}{Character. The indicator in a human readable format. Serves only to identify the indicator in the returned data.frame.}

\item{is_percentage}{Boolean. Indicates if expected values are represing percentages. Necessary to fix Value column because Quandl does not represents percentages property. Defaults to true.}
}
\value{
A data.frame containing:
\itemize{
\item Indicator (indicator name)
\item Date
\item StartDate
\item Year
\item Quarter
\item Month
\item Day
\item Weekday
\item Value
\item PctChange (based on Value)
}
}
\description{
Download daily and monthly values of several Brazilian economic indicators.
}
\examples{
\dontrun{
download_indicator_cdi_daily()
download_indicator_cdi_monhtly()
download_indicator_ibovespa_daily()}

}
\author{
Renato Dinhani
}
