% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stock_fundamentals.R
\name{download_stock_fundamentals}
\alias{download_stock_fundamentals}
\title{Download Stock Fundamentals}
\usage{
download_stock_fundamentals(ticker, folder = "temp")
}
\arguments{
\item{ticker}{Character. The ticker of a brazilian stock to download its fundamentals.}

\item{folder}{Character. A path to a folder where the downloaded data (ZIP file and Excel sheet) will be stored.}
}
\value{
A list containing:
\itemize{
\item ticker: Character. Ticker passed as parameter to this function.
\item filename: Character. Path to the downloaded Excel file containing the balance sheet and income statement.
\item balance: DataFrame. Balance sheet parsed from the downloaded Excel file.
\item income: DataFrame. Income statement parsed from the downloaded Excel file.
}
}
\description{
Download balance sheet and income statement from stocks listed in the Brazilian stock market.
}
\examples{
\dontrun{
download_stock_fundamentals("CIEL3")}

}
\author{
Renato Dinhani
}
