% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stocks_metadata.R
\name{download_stocks_metadata}
\alias{download_stocks_metadata}
\title{Download Stocks Metadata}
\usage{
download_stocks_metadata(folder = "temp/stocks/metadata/")
}
\arguments{
\item{folder}{Character. Path to a folder where the downloaded data (Excel sheet) will be stored.}
}
\value{
A data.frame containing:
\itemize{
\item Name
\item Ticker
\item FullName
\item MarketSegment
\item OrdinaryShares
\item PreferenceShares
\item TotalShares
}
}
\description{
Download information like ticker, number of shares and market segment from the companies listed in the Brazilian stock market.
}
\examples{
\dontrun{
download_stocks_metadata()}

}
\author{
Renato Dinhani
}
