% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stocks_prices.R
\name{download_stocks_prices}
\alias{download_stocks_prices}
\title{Download Stocks Prices}
\usage{
download_stocks_prices(tickers)
}
\arguments{
\item{tickers}{Character. Tickers of one or more brazilian stocks to download its daily prices.}
}
\value{
A data.frame containing:
\itemize{
\item Ticker
\item Date
\item StartDate
\item Open
\item High
\item Low
\item Close
\item Adjusted (Adjusted Close)
\item PctChange (based on Adjusted)
\item Volume
}
}
\description{
Download daily prices of stocks listed in the Brazilian stock market.
}
\examples{
\dontrun{
download_stocks_prices("CIEL3")}

}
\author{
Renato Dinhani
}
