% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadIndicatorCDI.R
\name{DownloadIndicatorCDIDaily}
\alias{DownloadIndicatorCDIDaily}
\alias{DownloadIndicatorCDIMonthly}
\title{Download Indicator - CDI}
\usage{
DownloadIndicatorCDIDaily()

DownloadIndicatorCDIMonthly()
}
\value{
A data.frame containing:
\itemize{
\item Indicator (indicator name)
\item Date
\item Value (percentage)
}
}
\description{
Download daily and monthly values of the CDI indicator.
}
\examples{
\dontrun{
DownloadIndicatorCDIDaily()
DownloadIndicatorCDIMonhtly()}

}
\author{
Renato Dinhani
}
