% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadIndicator.R
\name{DownloadIndicatorQuandl}
\alias{DownloadIndicatorQuandl}
\alias{DownloadIndicatorCDIDaily}
\alias{DownloadIndicatorCDIMonthly}
\alias{DownloadIndicatorIbovespaDaily}
\alias{DownloadIndicatorIpcaAccumulated}
\alias{DownloadIndicatorIpcaMonthly}
\alias{DownloadIndicatorSelicDaily}
\alias{DownloadIndicatorSelicMonthly}
\alias{DownloadIndicatorTjlpMonthly}
\title{Download Indicators}
\usage{
DownloadIndicatorQuandl(indicator.id, indicator.name = "")

DownloadIndicatorCDIDaily()

DownloadIndicatorCDIMonthly()

DownloadIndicatorIbovespaDaily()

DownloadIndicatorIpcaAccumulated()

DownloadIndicatorIpcaMonthly()

DownloadIndicatorSelicDaily()

DownloadIndicatorSelicMonthly()

DownloadIndicatorTjlpMonthly()
}
\arguments{
\item{indicator.id}{Chracter. The Quandl unique ID of the desired indicator.}

\item{indicator.name}{Character. The indicator in a human readable format. Serves only to identify the indicator in the returned data.frame.}
}
\value{
A data.frame containing:
\itemize{
\item Indicator (indicator name)
\item Date
\item Value (percentage)
}
}
\description{
Download daily and monthly values of several Brazilian economic indicators.
}
\examples{
\dontrun{
DownloadIndicatorCDIDaily()
DownloadIndicatorCDIMonhtly()
DownloadIndicatorIbovespaDaily()}

}
\author{
Renato Dinhani
}
