% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadStockPrices.R
\name{DownloadStockPrices}
\alias{DownloadStockPrices}
\title{Download Stock Prices}
\usage{
DownloadStockPrices(ticker)
}
\arguments{
\item{ticker}{Character. The ticker of a brazilian stock to download its daily prices.}
}
\value{
A data.frame containing:
\itemize{
\item Ticker
\item Date
\item Open
\item High
\item Low
\item Close
\item Adjusted (Adjusted Close)
\item Volume
}
}
\description{
Download daily prices of stocks listed in the Brazilian stock market.
}
\examples{
\dontrun{
DownloadStockPrices("CIEL3")}

}
\author{
Renato Dinhani
}
