% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stocks_fundamentals.R
\name{download_stocks_fundamentals}
\alias{download_stocks_fundamentals}
\title{Download Stocks Fundamentals}
\usage{
download_stocks_fundamentals(tickers,
  folder = "temp/stocks/fundamentals/")
}
\arguments{
\item{tickers}{Character. Tickers of one or more brazilian stocks to download its fundamentals.}

\item{folder}{Character. Path to a folder where the downloaded data (ZIP file and Excel sheet) will be stored.}
}
\value{
A data.frame containing data from the balance sheet and income statement.
}
\description{
Download balance sheet and income statement from stocks listed in the Brazilian stock market.
}
\examples{
\dontrun{
download_stocks_fundamentals("CIEL3")}

}
\author{
Renato Dinhani
}
